added users update logic
