+++ samples/json_probe.py
@@ -1,4 +1,5 @@
 import json
+import hashlib
 
 
 def main():
@@ -17,6 +18,9 @@ def main():
 
 def save(o):
     s = json.dumps(o, indent=2)
+    # m = hashlib.md5()
+    # m.update("000005fab4534d05api_key9a0554259914a86fb9e7eb014e4e5d52permswrite")
+    # m.hexdigest()
     with open('my.json', 'w') as f:
         f.write(s)
